Add polling and cachebuster to url
Browsers might not request a new build status when the response
is still in the browser cache. To prevent this a timestamp
is added to the url like ?1238763923, which will trigger a new
XHR request every time.

Additionally added an interval handler, to poll the server
every 60 seconds to decrease the chance of accidentally outdated
display of the build status.